Create MY LIFE
Life is a precious journey filled with experiences, both joyful and challenging, that shape who we are.in doing so I try to give shape to lives. so am by names of olea NSHUTI am a learner of Alx pathway program. I   finished secondary school in College  Appec Remera Rukoma